Added online changing of number of marks
When you set a new K (first accidents to search in query), you dont need to reload all requests, it will be done automatically. If there are no accidents in new first K, you will get an alert on your site.

Had to rewrite showAccidents, because async functions in JS are very hard.